pyproject.toml
[project]
name = "mlops"
version = "0.1.0"
description = "A project where we will be working on machine learning operations"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "uv>=0.9.22",
    "xgboost>=3.1.2",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
    "typer>=0.9.0",
    "matplotlib>=3.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=6.0.0",
    "isort>=5.13.0",
    "ruff>=0.8.0",
    "httpx>=0.27.0",
    "locust>=2.29.0",
    "google-api-python-client>=2.188.0",
    "google-cloud-storage>=2.10.0",
    "dvc-gs>=3.0.2",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
]
# "pre-commit = 4.0.0"


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", marker = "platform_system != 'Darwin'" },
]
torchvision = [
    { index = "pytorch-cpu", marker = "platform_system != 'Darwin'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[project.scripts]
make_data = "src.data.make_dataset:main"
train_rnn = "src.models.train_rnn:train"
rnn_data = "src.features.rnn_data:main"

[tool.uv]
package = true
required-environments = ["sys_platform == 'linux' and platform_machine == 'x86_64'"] # For Cloud Build


[tool.black]
line-length = 100

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
